- Change kdcom.sys to kdcom.dll. - Remove all kdcom exports and files from HAL and link with kdcom.dll instead. - Add support in freeldr for loading kdcom.dll (and theoretically any other import, such as the Code Integrity DLL for kernel-mode signing, but for now I've only tested this one). - Refactor some of the HAL loading code to make it generic for any image. - Load HAL manually when loading kernel, before processing imports, to make sure that HAL is index 1 in the loader modules and loads before KDCOM. - Fix some build problems.
svn path=/trunk/; revision=24780

+++ reactos/boot/freeldr/freeldr/reactos/setupldr.c
@@ -314,9 +314,9 @@ VOID RunLoader(VOID)
   LoaderBlock.PageDirectoryStart = (ULONG)&PageDirectoryStart;
   LoaderBlock.PageDirectoryEnd = (ULONG)&PageDirectoryEnd;
   LoaderBlock.BootDevice = 0xffffffff;
-  LoaderBlock.CommandLine = (unsigned long)reactos_kernel_cmdline;
+  LoaderBlock.CommandLine = reactos_kernel_cmdline;
   LoaderBlock.ModsCount = 0;
-  LoaderBlock.ModsAddr = (unsigned long)reactos_modules;
+  LoaderBlock.ModsAddr = reactos_modules;
   LoaderBlock.MmapLength = (unsigned long)MachGetMemoryMap((PBIOS_MEMORY_MAP)(PVOID)&reactos_memory_map, 32) * sizeof(memory_map_t);
   if (LoaderBlock.MmapLength)
     {
